refactor: update deployedOn descriptions for clarity
Updated the description for the deployedOn field across various schemas to specify the type of asset being deployed, enhancing clarity and consistency in the documentation.
