+++ kit/dapp/src/lib/queries/cryptocurrency/cryptocurrency-schema.ts
@@ -47,7 +47,7 @@ export const OnChainCryptoCurrencySchema = t.Object(
         "The percentage of total supply held by the top holders, indicating ownership concentration",
     }),
     deployedOn: t.StringifiedBigInt({
-      description: "The timestamp when the bond was deployed",
+      description: "The timestamp when the cryptocurrency was deployed",
     }),
   },
   {
