+++ kit/dapp/src/lib/queries/deposit/deposit-schema.ts
@@ -78,7 +78,7 @@ export const OnChainDepositSchema = t.Object(
         "The percentage of total supply held by the top holders, indicating ownership concentration",
     }),
     deployedOn: t.StringifiedBigInt({
-      description: "The timestamp when the bond was deployed",
+      description: "The timestamp when the deposit was deployed",
     }),
   },
   {
