+++ kit/dapp/src/lib/queries/equity/equity-schema.ts
@@ -64,7 +64,7 @@ export const OnChainEquitySchema = t.Object(
         "The percentage of total supply held by the top holders, indicating ownership concentration",
     }),
     deployedOn: t.StringifiedBigInt({
-      description: "The timestamp when the bond was deployed",
+      description: "The timestamp when the equity was deployed",
     }),
   },
   {
